SKALE-2115 Remove implicit calcultation of delegation state
