+++ .soliumrc.json
@@ -22,6 +22,7 @@
     "function-order": "error",
     "uppercase": "error",
     "error-reason": "error",
-    "security/no-inline-assembly": "warning"
+    "security/no-inline-assembly": "warning",
+    "security/no-block-members": "off"
   }
 }
